refactor: split for clarity
